change to greater precision for seconds_taken value
a job with many small tables would otherwise have many seconds=0

+++ squark-classic/config.py
@@ -42,7 +42,7 @@ ADMIN_LOAD_TIMINGTABLE_DDL="""CREATE TABLE {SCHEMA}.{TABLE} (
     build_number int,
     project_id varchar(50),
     table_name varchar(100),
-    seconds_taken int,
+    seconds_taken numeric(37,2),
     attempt_count int,
     source varchar(50),
     total_table_count int,
